feat : end of the security config
Former-commit-id: df526e6d86b48151c3cbb1a6b4765f1e27ef3f0d

+++ backend/src/main/java/com/ssafy/i5i/hotelAPI/domain/hotel/dto/request/AttractionCoordinateRequestDto.java
@@ -9,8 +9,8 @@ import java.math.BigDecimal;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 public class AttractionCoordinateRequestDto {
-    private BigDecimal latitude;
-    private BigDecimal longitude;
+    private Double latitude;
+    private Double longitude;
     private Long distance;
     private String sorted;
 }
